+++ PSC.Providers/CommonTables/AzureLocationProvider.cs
@@ -52,7 +52,7 @@ namespace PSC.Providers.CommonTables
             return value.ID;
         }
 
-        public async Task ReplaceAsync(int id, AzureLocation value)
+        public async Task ReplaceAsync(long id, AzureLocation value)
         {
             _db.Update(value);
             await _db.SaveChangesAsync();
