+++ PSC.Providers/CommonTables/AzureResourceProvider.cs
@@ -52,7 +52,7 @@ namespace PSC.Providers.CommonTables
             return value.ID;
         }
 
-        public async Task ReplaceAsync(int id, AzureResource value)
+        public async Task ReplaceAsync(long id, AzureResource value)
         {
             _db.Update(value);
             await _db.SaveChangesAsync();
